major: remove query engine python packages from requirement/base.txt
