+++ requirements/base.txt
@@ -41,25 +41,13 @@ markdown2
 # Utils
 typing-extensions==3.7.4.1
 
-
 # Optional Packages ----------
-# Querying
-# Hive/Presto
+# HMS
 sasl==0.2.1
 thrift==0.11.0
 thrift-sasl==0.3.0
-PyHive[hive, presto]==0.6.3
 kazoo==2.5.0
 hmsclient==0.1.1
-# Druid
-pydruid==0.5.7
-# BigQuery
-google-cloud-bigquery==1.28.0
-# Snowflake
-# snake-connector-python version needs to specified since it is a transitive dependency of snowflake-sqlalchemy
-# https://github.com/pinterest/querybook/pull/669
-snowflake-connector-python==2.6.1
-snowflake-sqlalchemy==1.2.4
 
 # Query Store
 # AWS Support
@@ -67,9 +55,6 @@ requests-aws4auth==0.9
 boto3==1.9.201
 boto==2.49.0
 moto==2.2.4
-
-# Trino Plugin
-trino==0.305.0
 
 # Google GCD
 google-cloud-storage==1.37.0
