+++ requirements/engines/bigquery.txt
@@ -1,0 +1,1 @@
+google-cloud-bigquery==1.28.0
