+++ requirements/engines/druid.txt
@@ -1,0 +1,1 @@
+pydruid==0.5.7
