+++ requirements/engines/hive.txt
@@ -1,0 +1,1 @@
+PyHive[hive]==0.6.3
