+++ requirements/engines/trino.txt
@@ -1,0 +1,1 @@
+trino==0.305.0
